extend the current mapping
The proposed mapping is a result of our experience while using OCR: due to wrong language settings some individual characters from strange languages are used and could be healed by this setting (which could be extended by further mapping ...).
The mapping of umlauts could be discussed, but should be consistent: Is the pronunciation essential, i.e. the usual habit? Or should the mapping the same as for other 'accents', i.e. just keeping the base character?