Fix splitting complicated page ranges like `A-1-A-7`
There was an attempt to fix this in 31a89bb however the outer `$from,
$to` regex was preventing the count regex from doing anything so the fix
didn't work.

To support this it counts single items as it did before (`[1-9]\d+`) and
then splits items with hyphens on the first hyphen (no matter what other
text they contain). Then if it contains multiple hyphens we split it
again on the 'middle' hyphen, assuming that is the best one.

This also splits the splitting algorithm into its own function
(`Pagerange::split_range`) which allows me to remove the one in
`HighwirePress` as now it can just use the general one so will change
with it if we fix any more of the page range splitting (although it
should really just be stored as two items in the database).

Actually fixes #60.