Add 'export_file_as_attachment' option to force exports to be downloaded
This uses the
['Content-Disposition'](https://developer.mozilla.org/en-US/docs/Web/HTTP/Reference/Headers/Content-Disposition)
request header to inform the browser that a file should be downloaded
rather than viewed inline.

By default with 'export_file_as_attachment' unset (or set to `0`) the
behaviour will remain the same as 'Content-Disposition' generally
defaults to inline (although I don't know that this is actually
specified in the spec). When it is set to `1` however the files
generated by 'Export' on searches/browse views (`cgi/exportview`) and
those generated by 'Export' on a single EPrint (`cgi/export`) will
inform the browser that it should always download them. This appears to
be generally followed across browsers although I have seen mentions that
some mobile browsers may disregard it.

Fixes #492.