Allow plugins to override the general export 'Content-Disposition'
Sometimes a plugin may want to open 'inline' even when the rest of them
are being opened as 'attachment' or vice versa for example if you had an
export plugin that produced a PDF you would likely want it to be
'inline'.

To support this there is a new `content_disposition` config field on
Export plugins that defaults to `undef`. This won't change anything
while unset however it can then be set to `'inline'` or `'attachment'`
which will force that 'Content-Disposition' no matter what the general
`export_file_as_attachment` config is set to.