Prevent double presses on search buttons sending multiple requests (#495)
If a user clicks the search button and then clicks it again because
they're not sure whether it went through they end up sending another
request. While it appears the browser will cancel the first request (as
a new one has taken priority) it may well be that this still dispatches
a variety of database queries - and there is no particular reason to
allow it.

This prevents double presses by disabling the search button after its
first click event (as recognised by `name: '_action_search'` which
appears to only be used by searches). There is then a bit of further
complexity to re-enable the button when we come back to the page as the
back button is complicated on a modern browser.