Mark the '<meta>' sections with comments (#496)
As PRed for `metatags` in eprints/metatags#4 this adds HTML comments to
the top of the 'Highwire Press' and 'PRISM' `<meta>` tag sections, as
well as adding them to the pre-existing 'Dublin Core' and 'EPrints'
tags.

While these tags are absolutely not designed for humans to read it feels
unnecessary to actively avoid human readability so this is a very small
change to help slightly if scanning through the tags (which also makes
it easier to see which tags a particular repository has enabled).