Apply 'export_file_as_attachment' to search
While I thought I had tested this when implementing this generally
(eprints/eprints3.4#493) it appears that I did so incorrectly (or not at
all) as while browse views (`exportview`) and individual EPrints
(`export`) work, searches would still use `inline` as that is the
default disposition.

This is a follow-on to eprints/eprints3.4@e3da74f, see that for more
detail on 'Content-Disposition'.