Add an export plugin for document URLs
This generates a simple HTML file containing links to all of the public
documents the given search contains. This should both be easier to
download them all by hand and very easy to parse with a script as it
would just have to look for `href="..."`.

It's probably worth considering whether this should add all documents
accessible to the current user rather than just those that are public as
if you are an admin you may be intentionally downloading private
documents.

Fixes #386.