Export links to all documents that the user can access
It would previously only include documents that are public to all users
however if you are an admin you may want to download documents that are
not publically available, and at the very least you want to know that it
has skipped documents. This just makes the documents it shows base
themselves on the users permissions rather than just document publicity.

It may be worth adding a bit of text (something like `(private)`) to
make it clear to a user that some of these documents are not publically
available, although the worst that can happen would probably be that
someone not logged in clicks a link and it tells them they need to log
in.