Add an export plugin for document URLs (#513)
This generates a simple HTML file containing links to all of the public documents the given search contains. This should both be easier to download them all by hand and very easy to parse with a script as it would just have to look for href="...".

This will show all documents that the user could access (so if an admin is logged in it would include all documents) rather than just those that are public. When there are private links in among them it will add (private) to make it clear that they shouldn't be shared to someone without admin permissions.

Fixes #386.