Add clear buttons to multiple 'Compound' fields (#512)
This is similar to the equivalent PR for 3.5 (eprints/eprints3.5#218)
however simpler. Because there is no contributors field it doesn't have
to handle clearing `span`s and it also doesn't handle defaults. This is
both because there are no core clearable defaults (as `contributors` is
the only one) and because that change ends up being a bit more involved
and is therefore better suited for 3.5.

Fixes #368.