Remove 'Document URLs' export from non-EPrint searches
This was originally based on the 'Object IDs' plugin which wants to show
up for every type of dataobject but doesn't want to show up for
individual EPrints. By contrast there is value in exporting the document
URLs for a single EPrint but no value in exporting them for say 'users'.

This also actually simplifies the exporter as it doesn't have to loop
through all the records or handle writing to passed in files (/STDOUT).