Remove the progress bar even if the upload fails
Previously the way upload ending worked was semi-converted from the old
'Fetch' solution so still had some quirks. In particular it worked by
RegEx parsing the returned javascript to get the `docid`. However if the
upload failed the `docid` would be set to 'null' which doesn't match
`[0-9]*` and it would therefore fail to remove the progress bar.

This replaces that madness with a sane solution that just sends JSON and
therefore still parses if the `docid` is missing it just doesn't refresh
the documents.

To help the user understand what happened (rather than the progress bar
just vanishing) it also now sends the Processor messages and adds them
to the correct part of the page. This allows it to at least bring up
'File upload failed.' which while not a very useful error message is
better than nothing.