Remove 'progressid' from file uploads
This was left over from when the progress bar was implemented in a
pretty hacky way and so it needed some way to track which progress bar a
particular request referred to.

As this was no longer actually used by standard file upload (it appears
to still be mentioned by Batch Edit?) it was just a matter of removing
anywhere that creates or passes it.