Show the correct timezone when `render_type => 'long'`
Previously it would always show 'UTC' however as we can now handle
different timezones it should also show different timezone strings.

Because of the use of `DateTime` this will change the string depending
on whether this date is within summer time or not, for example `GMT` or
`BST` with `Europe/London`.